Правка функции генерации данных и перенос getRandomNumber в mock
