+++ js/main.js
@@ -1,3 +1,5 @@
-import {photoDescriptions} from './mock/demo-data.js';
+import {createPhotoDescriptions} from './mock/demo-data.js';
 
-photoDescriptions;
+const NUMBER_PHOTO_DESCRIPTIONS = 25;
+
+createPhotoDescriptions(NUMBER_PHOTO_DESCRIPTIONS);
